+++ NEWS.md
@@ -1,7 +1,8 @@
-# sjmisc 1.6-1
+# sjmisc 1.6-2
 
 ## New functions
 
 * `merge_df` to fully join labelled data frame and preserve value and variable labels.
 * `wtd_sd` to compute weighted standard deviations.
 * `wtd_se` to compute weighted standard errors.
+* `get_note` and `set_note` to annotate vectors.